Agregado Fetch y Librerías

+++ js/carrito.js
@@ -71,6 +71,26 @@ function actualizarBtnEliminar() {
 }
 
 function eliminarDelCarrito(e) {
+
+  Toastify({
+    text: "Producto eliminado!",
+    duration: 3000,
+    close: true,
+    gravity: "top", // `top` or `bottom`
+    position: "right", // `left`, `center` or `right`
+    stopOnFocus: true, // Prevents dismissing of toast on hover
+    style: {
+      background: "linear-gradient(to right, #AA5042, #af675c)",
+      borderRadius: "1.5rem",
+      fontSize: ".75rem"
+    },
+    offset: {
+        x: "1.5rem", // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+        y: "1.5rem" // vertical axis - can be a number or a string indicating unity. eg: '2em'
+      },
+    onClick: function(){} // Callback after click
+  }).showToast();
+
   const idBoton = e.currentTarget.id;
   const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
 
@@ -86,9 +106,25 @@ btnVaciar.addEventListener("click", vaciarCarrito);
 
 function vaciarCarrito(){
 
-  productosEnCarrito.length = 0;
-  localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-  cargarProductosCarrito();
+  Swal.fire({
+    title: '¿Estás seguro?',
+    icon: 'question',
+    iconColor: '#AA5042',
+    html:
+      `Se van a eliminar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+    showCancelButton: true,
+    confirmButtonText:
+      'Si',
+    confirmButtonColor: '#af675c',
+    cancelButtonText:
+      'No',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      productosEnCarrito.length = 0;
+      localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+      cargarProductosCarrito();
+    }
+  })
 
 }
